Move checkbox related code to checkbox
