+++ js/filters/components/checkbox.js
@@ -1,3 +1,5 @@
+import { createSubContainer, createCollectionSubtitle } from "./lib.js";
+
 const createCheckboxItemContainer = () => {
   const container = document.createElement("div");
   container.classList.add("checkbox-title-container");
@@ -99,4 +101,80 @@ export function filterOutByNormalCheckboxes(object, checkboxGroup) {
   });
 
   return satisfies;
+}
+
+export function createCheckboxContainer() {
+  const container = document.createElement("div");
+  container.classList.add("filter-checkbox-container");
+  return container;
+}
+
+export function createButtonForCheckboxes(text) {
+  const button = document.createElement("button");
+  button.classList.add("checkbox-button");
+  button.innerText = text;
+  return button;
+}
+
+export function buildEnumCheckboxes(label, propertyName, values) {
+  const groupContainer = createSubContainer();
+  groupContainer.appendChild(createCollectionSubtitle(label));
+  const checkboxesContainer = createCheckboxContainer();
+
+  const checkboxes = [...values].map((value) => {
+    const checkbox = new CheckboxComponent(propertyName, value, value, true);
+    checkboxesContainer.appendChild(checkbox.render());
+    checkbox.checked(true);
+    return checkbox;
+  });
+
+  groupContainer.appendChild(checkboxesContainer);
+  return { groupContainer, checkboxes };
+}
+
+export function buildBitfieldCheckboxes(collection) {
+  const container = createSubContainer();
+  const div = document.createElement("div");
+  div.classList.add("collection-checkboxes-handler");
+  const title = createCollectionSubtitle("Collection");
+  const buttonsDiv = document.createElement("div");
+  const selectAll = createButtonForCheckboxes("Select all");
+  const clearAll = createButtonForCheckboxes("Clear all");
+  div.appendChild(title);
+  buttonsDiv.appendChild(selectAll);
+  buttonsDiv.appendChild(clearAll);
+  div.appendChild(buttonsDiv);
+  container.appendChild(div);
+  const checkboxesContainer = createCheckboxContainer();
+
+  const checkboxes = [];
+  const collections = new Set();
+  collection.forEach((object) => collections.add(object.collectionName));
+
+  collections.forEach((collectionName) => {
+    const checkbox = new CheckboxComponent(
+      "collectionName",
+      collectionName,
+      collectionName,
+      true,
+    );
+    checkboxes.push(checkbox);
+    checkboxesContainer.appendChild(checkbox.render());
+    checkbox.checked(true);
+  });
+  container.appendChild(checkboxesContainer);
+
+  selectAll.addEventListener("click", () => {
+    checkboxes.forEach((checkbox) => {
+      checkbox.checked(true);
+    });
+  });
+
+  clearAll.addEventListener("click", () => {
+    checkboxes.forEach((checkbox) => {
+      checkbox.checked(false);
+    });
+  });
+
+  return [container, checkboxes];
 }
